New, simplier design.
First, limit delegations to interfaces and protocols and never classes.
Second, change the way the monoid? multimethod dispatches to include the method
name explicitly. Third, no longer walk through a tree of interfaces but instead
rely the root interface to know all of its methods.

The effect of the above is that an interface can now override the monoid?
multimethod of higher values.